Refactor filter packages
